media-player/requirements.txt
fastapi==0.95.1
pymongo==4.3.3
uvicorn==0.21.1
google-cloud-storage==2.8.0
python-multipart==0.0.6
Jinja2==3.1.2